:lipstick: Redesign News View
The news view got completely redesigned.
We now do use a Wrap instead of GridView. So that there are the right count of articles
shown on the same time
on every (smaller) screen size. (e.g. Smartphone 1, Small tablet 2, Big tablet 3).
Also CachedNetworkImage got updated. This allows us to just use cachednetworkimage for
the images of the articles, before that did not work on web