:sparkles: Add linux support
Add support for building the app as an linux
executable. This includes the files necessary for flutter,
a github workflow for betaBuild and necessary changes to the
App to have it ignore linux in some parts.

That way PushNotifications and Webmail do not exist/work on linux
this is expected and accepted behaviour.